Refactor : TechStackService 기능을 AdminService 기능으로 옮김
