+++ src/main/java/com/example/feedprep/domain/user/service/AdminServiceImpl.java
@@ -2,6 +2,9 @@ package com.example.feedprep.domain.user.service;
 
 import com.example.feedprep.common.exception.base.CustomException;
 import com.example.feedprep.common.exception.enums.ErrorCode;
+import com.example.feedprep.domain.techstack.dto.CreateTechStackRequestDto;
+import com.example.feedprep.domain.techstack.entity.TechStack;
+import com.example.feedprep.domain.techstack.repository.TechStackRepository;
 import com.example.feedprep.domain.user.dto.response.TutorResponseDto;
 import com.example.feedprep.domain.user.entity.User;
 import com.example.feedprep.domain.user.enums.UserRole;
@@ -15,6 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class AdminServiceImpl implements AdminService{
 
     private final UserRepository userRepository;
+    private final TechStackRepository techStackRepository;
 
     @Override
     @Transactional
@@ -29,5 +33,18 @@ public class AdminServiceImpl implements AdminService{
         user.setRole(UserRole.APPROVED_TUTOR);
 
         return new TutorResponseDto(user.getRole());
+    }
+
+    @Override
+    public void createTechStack(CreateTechStackRequestDto requestDto) {
+        TechStack techStack = new TechStack(requestDto.getTechStack());
+        techStackRepository.save(techStack);
+    }
+
+    @Transactional
+    @Override
+    public void deleteTechStack(Long techId) {
+        TechStack techStack = techStackRepository.findByIdOrElseThrow(techId);
+        techStackRepository.delete(techStack);
     }
 }
