+++ src/test/java/com/example/feedprep/domain/user/service/AdminServiceImplTest.java
@@ -1,0 +1,20 @@
+package com.example.feedprep.domain.user.service;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class AdminServiceImplTest {
+
+    @Test
+    void approveTutor() {
+    }
+
+    @Test
+    void createTechStack() {
+    }
+
+    @Test
+    void deleteTechStack() {
+    }
+}
